+++ src/components/borderBox13/index.js
@@ -11,47 +11,55 @@ import useAutoResize from '../../use/autoResize'
 
 import './style.less'
 
-const defaultColor = ['#2862b7', '#2862b7']
+const defaultColor = ['#6586ec', '#2cf7fe']
 
 const BorderBox = ({ children, className, style, color = [] }) => {
   const { width, height, domRef } = useAutoResize()
 
   const mergedColor = useMemo(() => deepMerge(deepClone(defaultColor, true), color || []), [color])
 
-  const classNames = useMemo(() => classnames('dv-border-box-3', className), [
+  const classNames = useMemo(() => classnames('dv-border-box-13', className), [
     className
   ])
 
   return (
     <div className={classNames} style={style} ref={domRef}>
       <svg className='dv-border-svg-container' width={width} height={height}>
-        <polyline
-          className='dv-bb3-line1'
+        <path
+          fill='transparent'
           stroke={mergedColor[0]}
-          points={`4, 4 ${width - 22} ,4 ${width - 22}, ${height -
-            22} 4, ${height - 22} 4, 4`}
+          d={`
+            M 5 20 L 5 10 L 12 3  L 60 3 L 68 10
+            L ${width - 20} 10 L ${width - 5} 25
+            L ${width - 5} ${height - 5} L 20 ${height - 5}
+            L 5 ${height - 20} L 5 20
+          `}
         />
-        <polyline
-          className='dv-bb3-line2'
-          stroke={mergedColor[1]}
-          points={`10, 10 ${width - 16}, 10 ${width - 16}, ${height -
-            16} 10, ${height - 16} 10, 10`}
+
+        <path
+          fill='transparent'
+          strokeWidth='3'
+          strokeLinecap='round'
+          strokeDasharray='10, 5'
+          stroke={mergedColor[0]}
+          d='M 16 9 L 61 9'
         />
-        <polyline
-          className='dv-bb3-line2'
+
+        <path
+          fill='transparent'
           stroke={mergedColor[1]}
-          points={`16, 16 ${width - 10}, 16 ${width - 10}, ${height -
-            10} 16, ${height - 10} 16, 16`}
+          d='M 5 20 L 5 10 L 12 3  L 60 3 L 68 10'
         />
-        <polyline
-          className='dv-bb3-line2'
+
+        <path
+          fill='transparent'
           stroke={mergedColor[1]}
-          points={`22, 22 ${width - 4}, 22 ${width - 4}, ${height -
-            4} 22, ${height - 4} 22, 22`}
+          d={`M ${width - 5} ${height - 30} L ${width - 5} ${height - 5} L ${width - 30} ${height - 5}`}
         />
       </svg>
-
-      <div className='border-box-content'>{children}</div>
+      <div className='border-box-content'>
+        {children}
+      </div>
     </div>
   )
 }