+++ src/components/borderBox2/index.js
@@ -4,12 +4,19 @@ import PropTypes from 'prop-types'
 
 import classnames from 'classnames'
 
+import { deepMerge } from '@jiaminghi/charts/lib/util/index'
+import { deepClone } from '@jiaminghi/c-render/lib/plugin/util'
+
 import useAutoResize from '../../use/autoResize'
 
 import './style.less'
 
-const BorderBox = ({ children, className, style }) => {
+const defaultColor = ['#fff', 'rgba(255, 255, 255, 0.6)']
+
+const BorderBox = ({ children, className, style, color = [] }) => {
   const { width, height, domRef } = useAutoResize()
+
+  const mergedColor = useMemo(() => deepMerge(deepClone(defaultColor, true), color || []), [color])
 
   const classNames = useMemo(() => classnames('dv-border-box-2', className), [
     className
@@ -19,19 +26,19 @@ const BorderBox = ({ children, className, style }) => {
     <div className={classNames} style={style} ref={domRef}>
       <svg className='dv-border-svg-container' width={width} height={height}>
         <polyline
-          className='dv-bb2-line1'
+          stroke={mergedColor[0]}
           points={`2, 2 ${width - 2} ,2 ${width - 2}, ${height -
             2} 2, ${height - 2} 2, 2`}
         />
         <polyline
-          className='dv-bb2-line2'
+          stroke={mergedColor[1]}
           points={`6, 6 ${width - 6}, 6 ${width - 6}, ${height -
             6} 6, ${height - 6} 6, 6`}
         />
-        <circle cx='11' cy='11' r='1' />
-        <circle cx={width - 11} cy='11' r='1' />
-        <circle cx={width - 11} cy={height - 11} r='1' />
-        <circle cx='11' cy={height - 11} r='1' />
+        <circle fill={mergedColor[0]} cx='11' cy='11' r='1' />
+        <circle fill={mergedColor[0]} cx={width - 11} cy='11' r='1' />
+        <circle fill={mergedColor[0]} cx={width - 11} cy={height - 11} r='1' />
+        <circle fill={mergedColor[0]} cx='11' cy={height - 11} r='1' />
       </svg>
       <div className='border-box-content'>{children}</div>
     </div>
@@ -41,7 +48,8 @@ const BorderBox = ({ children, className, style }) => {
 BorderBox.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
-  style: PropTypes.object
+  style: PropTypes.object,
+  color: PropTypes.array
 }
 
 export default BorderBox