+++ src/components/borderBox6/index.js
@@ -4,12 +4,19 @@ import PropTypes from 'prop-types'
 
 import classnames from 'classnames'
 
+import { deepMerge } from '@jiaminghi/charts/lib/util/index'
+import { deepClone } from '@jiaminghi/c-render/lib/plugin/util'
+
 import useAutoResize from '../../use/autoResize'
 
 import './style.less'
 
-const BorderBox = ({ children, className, style }) => {
+const defaultColor = ['rgba(255, 255, 255, 0.35)', 'gray']
+
+const BorderBox = ({ children, className, style, color = [] }) => {
   const { width, height, domRef } = useAutoResize()
+
+  const mergedColor = useMemo(() => deepMerge(deepClone(defaultColor, true), color || []), [color])
 
   const classNames = useMemo(() => classnames('dv-border-box-6', className), [
     className
@@ -18,24 +25,26 @@ const BorderBox = ({ children, className, style }) => {
   return (
     <div className={classNames} style={style} ref={domRef}>
       <svg className='dv-svg-container' width={width} height={height}>
-        <circle cx='5' cy='5' r='2' />
-        <circle cx={width - 5} cy='5' r='2' />
-        <circle cx={width - 5} cy={height - 5} r='2' />
-        <circle cx='5' cy={height - 5} r='2' />
-        <polyline points={`10, 4 ${width - 10}, 4`} />
-        <polyline points={`10, ${height - 4} ${width - 10}, ${height - 4}`} />
-        <polyline points={`5, 70 5, ${height - 70}`} />
-        <polyline points={`${width - 5}, 70 ${width - 5}, ${height - 70}`} />
-        <polyline points={`3, 10, 3, 50`} />
-        <polyline points={`7, 30 7, 80`} />
-        <polyline points={`${width - 3}, 10 ${width - 3}, 50`} />
-        <polyline points={`${width - 7}, 30 ${width - 7}, 80`} />
-        <polyline points={`3, ${height - 10} 3, ${height - 50}`} />
-        <polyline points={`7, ${height - 30} 7, ${height - 80}`} />
+        <circle fill={mergedColor[1]} cx='5' cy='5' r='2' />
+        <circle fill={mergedColor[1]} cx={width - 5} cy='5' r='2' />
+        <circle fill={mergedColor[1]} cx={width - 5} cy={height - 5} r='2' />
+        <circle fill={mergedColor[1]} cx='5' cy={height - 5} r='2' />
+        <polyline fill={mergedColor[0]} points={`10, 4 ${width - 10}, 4`} />
+        <polyline fill={mergedColor[0]} points={`10, ${height - 4} ${width - 10}, ${height - 4}`} />
+        <polyline fill={mergedColor[0]} points={`5, 70 5, ${height - 70}`} />
+        <polyline fill={mergedColor[0]} points={`${width - 5}, 70 ${width - 5}, ${height - 70}`} />
+        <polyline fill={mergedColor[0]} points={`3, 10, 3, 50`} />
+        <polyline fill={mergedColor[0]} points={`7, 30 7, 80`} />
+        <polyline fill={mergedColor[0]} points={`${width - 3}, 10 ${width - 3}, 50`} />
+        <polyline fill={mergedColor[0]} points={`${width - 7}, 30 ${width - 7}, 80`} />
+        <polyline fill={mergedColor[0]} points={`3, ${height - 10} 3, ${height - 50}`} />
+        <polyline fill={mergedColor[0]} points={`7, ${height - 30} 7, ${height - 80}`} />
         <polyline
+          fill={mergedColor[0]}
           points={`${width - 3}, ${height - 10} ${width - 3}, ${height - 50}`}
         />
         <polyline
+          fill={mergedColor[0]}
           points={`${width - 7}, ${height - 30} ${width - 7}, ${height - 80}`}
         />
       </svg>
@@ -48,7 +57,8 @@ const BorderBox = ({ children, className, style }) => {
 BorderBox.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
-  style: PropTypes.object
+  style: PropTypes.object,
+  color: PropTypes.array
 }
 
 export default BorderBox