+++ src/components/decoration9/index.js
@@ -4,16 +4,26 @@ import PropTypes from 'prop-types'
 
 import classnames from 'classnames'
 
+import { fade } from '@jiaminghi/color'
+
+import { deepMerge } from '@jiaminghi/charts/lib/util'
+
+import { deepClone } from '@jiaminghi/c-render/lib/plugin/util'
+
 import useAutoResize from '../../use/autoResize'
 
 import './style.less'
 
+const defaultColor = ['rgba(3, 166, 224, 0.8)', 'rgba(3, 166, 224, 0.5)']
+
 const svgWH = [100, 100]
 
-const Decoration = ({ children, className, style }) => {
+const Decoration = ({ children, className, style, color = [], dur = 3 }) => {
   const { width, height, domRef } = useAutoResize()
 
   const polygonIdRef = useRef(`decoration-9-polygon-${Date.now()}`)
+
+  const mergedColor = useMemo(() => deepMerge(deepClone(defaultColor, true), color || []), [color])
 
   const svgScale = useMemo(() => {
     const [w, h] = svgWH
@@ -44,7 +54,7 @@ const Decoration = ({ children, className, style }) => {
           cy='50'
           r='45'
           fill='transparent'
-          stroke='rgba(3, 166, 224, 0.5)'
+          stroke={mergedColor[1]}
           strokeWidth='10'
           strokeDasharray='80, 100, 30, 100'
         >
@@ -52,7 +62,7 @@ const Decoration = ({ children, className, style }) => {
             attributeName='transform'
             type='rotate'
             values='0 50 50;360 50 50'
-            dur='3s'
+            dur={`${dur}s`}
             repeatCount='indefinite'
           />
         </circle>
@@ -62,7 +72,7 @@ const Decoration = ({ children, className, style }) => {
           cy='50'
           r='45'
           fill='transparent'
-          stroke='rgba(3, 166, 224, 0.8)'
+          stroke={mergedColor[0]}
           strokeWidth='6'
           strokeDasharray='50, 66, 100, 66'
         >
@@ -70,7 +80,7 @@ const Decoration = ({ children, className, style }) => {
             attributeName='transform'
             type='rotate'
             values='0 50 50;-360 50 50'
-            dur='3s'
+            dur={`${dur}s`}
             repeatCount='indefinite'
           />
         </circle>
@@ -80,7 +90,7 @@ const Decoration = ({ children, className, style }) => {
           cy='50'
           r='38'
           fill='transparent'
-          stroke='rgba(3, 166, 224, 0.2)'
+          stroke={fade(mergedColor[1] || defaultColor[1], 30)}
           strokeWidth='1'
           strokeDasharray='5, 1'
         />
@@ -88,17 +98,17 @@ const Decoration = ({ children, className, style }) => {
           <use
             key={i}
             href={`#${polygonIdRef.current}`}
-            stroke='rgba(3, 166, 224, 0.6)'
+            stroke={mergedColor[1]}
             fill={
-              Math.random() > 0.4 ? 'transparent' : 'rgba(3, 166, 224, 0.8)'
+              Math.random() > 0.4 ? 'transparent' : mergedColor[0]
             }
           >
             <animateTransform
               attributeName='transform'
               type='rotate'
               values='0 50 50;360 50 50'
-              dur='3s'
-              begin={`${i * 0.15}s`}
+              dur={`${dur}s`}
+              begin={`${i * dur / 20}s`}
               repeatCount='indefinite'
             />
           </use>
@@ -109,7 +119,7 @@ const Decoration = ({ children, className, style }) => {
           cy='50'
           r='26'
           fill='transparent'
-          stroke='rgba(3, 166, 224, 0.2)'
+          stroke={fade(mergedColor[1] || defaultColor[1], 30)}
           strokeWidth='1'
           strokeDasharray='5, 1'
         />
@@ -123,7 +133,9 @@ const Decoration = ({ children, className, style }) => {
 Decoration.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string,
-  style: PropTypes.object
+  style: PropTypes.object,
+  color: PropTypes.array,
+  dur: PropTypes.number
 }
 
 export default Decoration